+++ src/main/java/insane96mcp/iguanatweaksreborn/mixin/FireBlockMixin.java
@@ -1,18 +1,22 @@
 package insane96mcp.iguanatweaksreborn.mixin;
 
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
 import insane96mcp.iguanatweaksreborn.module.world.coalfire.CoalFire;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
+import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.FireBlock;
+import net.minecraftforge.server.ServerLifecycleHooks;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(FireBlock.class)
 public class FireBlockMixin {
-	@Inject(at = @At("RETURN"), method = "getFireTickDelay", cancellable = true)
-	private static void setFireTickDelay(RandomSource p_221149_, CallbackInfoReturnable<Integer> cir) {
-		if (CoalFire.changeFireSpreadSpeed())
-			cir.setReturnValue((int) (cir.getReturnValue() / CoalFire.fireSpreadSpeedMultiplier));
+	@ModifyReturnValue(at = @At("RETURN"), method = "getFireTickDelay")
+	private static int setFireTickDelay(int original, RandomSource random) {
+		ServerLevel level = ServerLifecycleHooks.getCurrentServer().getLevel(Level.OVERWORLD);
+		if (level == null)
+			return original;
+		return original / level.getGameRules().getInt(CoalFire.RULE_FIRESPEEDMULTIPLIER);
 	}
 }