+++ src/main/java/insane96mcp/iguanatweaksreborn/module/world/coalfire/CoalFire.java
@@ -6,14 +6,17 @@ import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import insane96mcp.insanelib.base.config.LoadFeature;
+import net.minecraft.world.level.GameRules;
 
 @Label(name = "Coal & Fire")
 @LoadFeature(module = Modules.Ids.WORLD)
 public class CoalFire extends Feature {
 
-    @Config(min = 0d, max = 100)
+    public static final GameRules.Key<GameRules.IntegerValue> RULE_FIRESPEEDMULTIPLIER = GameRules.register("iguanatweaks:fireSpeedMultiplier", GameRules.Category.UPDATES, GameRules.IntegerValue.create(4));
+
+    /*@Config(min = 0d, max = 100)
     @Label(name = "Fire spread speed multiplier", description = "How much faster fire ticks and spreads.")
-    public static Double fireSpreadSpeedMultiplier = 4d;
+    public static Double fireSpreadSpeedMultiplier = 4d;*/
 
     @Config
     @Label(name = "Unlit campfire", description = "If true, campfires must be lit")
@@ -21,10 +24,6 @@ public class CoalFire extends Feature {
 
     public CoalFire(Module module, boolean enabledByDefault, boolean canBeDisabled) {
         super(module, enabledByDefault, canBeDisabled);
-    }
-
-    public static boolean changeFireSpreadSpeed() {
-        return Feature.isEnabled(CoalFire.class) && fireSpreadSpeedMultiplier != 1d;
     }
 
     public static boolean areCampfiresUnlit() {